# Use platform-specific methods for FileAccess reading and writing