--- a/drivers/unix/file_access_unix.cpp
+++ b/drivers/unix/file_access_unix.cpp
@@ -228,6 +228,51 @@ uint8_t FileAccessUnix::get_8() const {
 	return b;
 }
 
+uint16_t FileAccessUnix::get_16() const {
+	ERR_FAIL_NULL_V_MSG(f, 0, "File must be opened before use.");
+
+	uint16_t b = 0;
+	if (fread(&b, 1, 2, f) != 2) {
+		check_errors();
+	}
+
+	if (big_endian) {
+		b = BSWAP16(b);
+	}
+
+	return b;
+}
+
+uint32_t FileAccessUnix::get_32() const {
+	ERR_FAIL_NULL_V_MSG(f, 0, "File must be opened before use.");
+
+	uint32_t b = 0;
+	if (fread(&b, 1, 4, f) != 4) {
+		check_errors();
+	}
+
+	if (big_endian) {
+		b = BSWAP32(b);
+	}
+
+	return b;
+}
+
+uint64_t FileAccessUnix::get_64() const {
+	ERR_FAIL_NULL_V_MSG(f, 0, "File must be opened before use.");
+
+	uint64_t b = 0;
+	if (fread(&b, 1, 8, f) != 8) {
+		check_errors();
+	}
+
+	if (big_endian) {
+		b = BSWAP64(b);
+	}
+
+	return b;
+}
+
 uint64_t FileAccessUnix::get_buffer(uint8_t *p_dst, uint64_t p_length) const {
 	ERR_FAIL_COND_V(!p_dst && p_length > 0, -1);
 	ERR_FAIL_NULL_V_MSG(f, -1, "File must be opened before use.");
@@ -249,6 +294,36 @@ void FileAccessUnix::flush() {
 void FileAccessUnix::store_8(uint8_t p_dest) {
 	ERR_FAIL_NULL_MSG(f, "File must be opened before use.");
 	ERR_FAIL_COND(fwrite(&p_dest, 1, 1, f) != 1);
+}
+
+void FileAccessUnix::store_16(uint16_t p_dest) {
+	ERR_FAIL_NULL_MSG(f, "File must be opened before use.");
+
+	if (big_endian) {
+		p_dest = BSWAP16(p_dest);
+	}
+
+	ERR_FAIL_COND(fwrite(&p_dest, 1, 2, f) != 2);
+}
+
+void FileAccessUnix::store_32(uint32_t p_dest) {
+	ERR_FAIL_NULL_MSG(f, "File must be opened before use.");
+
+	if (big_endian) {
+		p_dest = BSWAP32(p_dest);
+	}
+
+	ERR_FAIL_COND(fwrite(&p_dest, 1, 4, f) != 4);
+}
+
+void FileAccessUnix::store_64(uint64_t p_dest) {
+	ERR_FAIL_NULL_MSG(f, "File must be opened before use.");
+
+	if (big_endian) {
+		p_dest = BSWAP64(p_dest);
+	}
+
+	ERR_FAIL_COND(fwrite(&p_dest, 1, 8, f) != 8);
 }
 
 void FileAccessUnix::store_buffer(const uint8_t *p_src, uint64_t p_length) {