Implements platform-specific variants of FileAccess's `get_16`, `get_32`, `get_64`, `store_16`, `store_32`, and `store_64`. Also adds MSVC-specific versions of `BSWAPxx` functions.

This results in a significant performance gain when reading or writing large amounts of binary data: ~5x faster writing and ~6x faster reading.

Benchmark results for writing data to a file: 

Windows x64, build - production
| data | master | PR |
| ------------- | ------------- | ------------- |
| 10000000 x uint64_t (78 mb) | 1.70s | 0.30s  |
| 50000000 x uint64_t (390 mb)  | 8.70s  | 1.80s |
| 10000 x uint64_t (79 kb)  | 0.0020s  | 0.0007s |
| 500000 x uint64_t big endian (4 mb)  | 0.085s  | 0.015s |
| 500000 x uint64_t little endian (4 mb)  | 0.086s  | 0.015s |

A small project for testing whether the changes work correctly: [file_access_test.zip](https://github.com/godotengine/godot/files/13219041/file_access_test.zip)

Progress:
- [x] FileAccessWindows
- [x] FileAccessUnix
- [x] FileAccessAndroid
- [x] Minimal testing project